cofi-mediator-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
docker>=7.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=23.0.0
